+++ R/utils.R
@@ -6,7 +6,7 @@
 #' @import survival
 #' @importFrom utils read.csv write.csv head
 #' @importFrom tidyr pivot_wider pivot_longer
-#' @importFrom dplyr select mutate all_of group_by summarise ungroup reframe ntile n across rename
+#' @importFrom dplyr select mutate all_of group_by summarise ungroup reframe ntile n across rename sym
 #' @importFrom fBasics shapiroTest lillieTest adTest jarqueberaTest sfTest
 NULL